+++ commas/src/main/java/tight/commas/domain/drink/api/DrinkApi.java
@@ -5,7 +5,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.web.util.UriComponentsBuilder;
 import reactor.core.publisher.Mono;
-import tight.commas.utils.WebClientUtils;
+import tight.commas.global.utils.WebClientUtils;
 import tight.commas.domain.drink.dto.DrinkDto;
 
 import java.util.List;
